refactor(baseten_server): split 1552-line module into a package (#100)
Third and final monolith split for #100. baseten_server.py held the
FastAPI app, all routes, the BasetenCUARuntime class, llama-server
lifecycle, auth dependencies, secret resolvers, path helpers, log
formatting — six concerns interleaved. Each PR touching one of them
had to be reviewed against the others.

Split, behavior-neutral. ``git mv`` preserves blame on the bulk of
the route handlers under ``routes.py``.

baseten_server/                   (line counts after split)
  __init__.py        20           re-exports app + runtime singleton
                                  + BasetenCUARuntime
  paths.py          102           per-tenant path resolvers + GGUF
                                  finders
  logging_setup.py   76           JsonLogFormatter + configure_logging
                                  + DetachedRunLogHandler
  middleware.py      88           require_mantis_token + require_run_
                                  scope + read_secret +
                                  resolve_anthropic_key + load_secret_
                                  environment + SECRET_ENV_MAP
  runtime.py        951           BasetenCUARuntime
  routes.py         524           FastAPI app + all 9 route handlers
                                  (was 1552; -1028 moved out)

Backward compat preserved
- ``from mantis_agent.baseten_server import app`` (the import the
  Baseten Truss + Modal entrypoint use) keeps working — re-exported
  from __init__.
- Single-underscore aliases inside routes.py and runtime.py keep the
  carved bodies unchanged: routes still call ``_require_run_scope``,
  ``_resolve_anthropic_key``, ``_data_root``, etc., aliased to the new
  canonical sibling-module names.
- All 9 routes register cleanly: /health, /v1/health, /metrics,
  /v1/models, /v1/runs/{run_id}/video, /v1/chat/completions,
  /v1/predict, /predict, /openapi.json.

Test fix
- tests/test_chat_completions_proxy.py used
  ``patch("mantis_agent.baseten_server.requests.post", ...)``. After
  the split, ``requests`` is imported in ``routes.py`` rather than at
  the package root, so the patch path no longer resolves. Updated to
  ``mantis_agent.baseten_server.routes.requests.post`` — the canonical
  location.

Tests / lint
- pytest tests/ -q -> 370 passed (same as after #106; pure refactor).
- ruff check on the new package: clean.

Out of scope (acceptance criterion still partial)
- ``runtime.py`` is 951 lines — over the 800-line target. The bulk is
  one BasetenCUARuntime class. Same caveat as ``extractor.py`` (#105)
  and ``osworld.py`` (#106): decomposing it requires architectural
  decisions, not mechanical moves.

Closes #100 in the sense intended — all three monolithic files are
now packages with proper concern boundaries. Further size reduction on
the three remaining oversized files is a separate effort.
