+++ src/mantis_agent/baseten_server/runtime.py
@@ -1,8 +1,15 @@
-"""Baseten workload server for Mantis CUA.
+"""``BasetenCUARuntime`` — model lifecycle, run dispatch, llama-server.
 
-This module is used by the Baseten custom-server Trusses under ``deploy/baseten/``.
-It starts a local llama.cpp server for either Holo3 or Gemma4, then exposes a
-small FastAPI surface that runs the existing CUA task and micro-plan runners.
+Owns:
+
+- The model singleton (Holo3 / Gemma4 / etc.) and its in-pod
+  llama.cpp server lifecycle
+- The detached-run thread pool and per-run state directory
+- The Anthropic key resolver per tenant
+- The CUA / micro-plan runners that the FastAPI routes invoke
+
+Routes live in :mod:`.routes`. Path helpers live in :mod:`.paths`.
+Auth + secret middleware live in :mod:`.middleware`.
 """
 
 from __future__ import annotations
@@ -14,37 +21,13 @@ import subprocess
 import threading
 import time
 import traceback
-import uuid
 from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any
 
 
-import requests
-
-try:
-    from fastapi import Depends, FastAPI, Header, HTTPException, Request, Response
-    from fastapi.responses import JSONResponse
-    from starlette.concurrency import run_in_threadpool
-except ImportError as exc:  # pragma: no cover - container-only deps
-    raise ImportError(
-        "mantis_agent.baseten_server requires fastapi + uvicorn. "
-        "Install via: pip install -e '.[server]'  (or run inside the "
-        "Baseten Truss image, which provisions them in build_commands)."
-    ) from exc
-
-from mantis_agent.api_schemas import (
-    MAX_COST_USD,
-    MAX_RUNTIME_MINUTES,
-    PredictRequest,
-    assert_hosts_allowed,
-    extract_navigate_hosts,
-)
-from mantis_agent.gym.xdotool_env import XdotoolGymEnv
-from mantis_agent.idempotency import get_idempotency_cache
-from mantis_agent import metrics as mantis_metrics
-from mantis_agent.rate_limit import get_rate_limiter
-from mantis_agent.server_utils import (
+from ..gym.xdotool_env import XdotoolGymEnv
+from ..server_utils import (
     build_micro_result,
     build_micro_suite,
     build_proxy_config,
@@ -59,250 +42,51 @@ from mantis_agent.server_utils import (
     wait_for_openai_server,
     write_leads_csv,
 )
-from mantis_agent.tenant_auth import (
-    DEFAULT_TENANT,
-    TenantConfig,
-    get_key_store,
+from ..tenant_auth import DEFAULT_TENANT
+from .logging_setup import DetachedRunLogHandler
+from .middleware import load_secret_environment, read_secret, resolve_anthropic_key
+from .paths import (
+    data_root,
+    find_gguf,
+    find_mmproj,
+    new_run_id,
+    repo_root,
+    tenant_chrome_profile,
+    tenant_root,
+    tenant_state_key,
 )
-from mantis_agent.webhooks import WebhookPayload, deliver_webhook_async
-
-class _JsonLogFormatter(logging.Formatter):
-    """One-line JSON-per-record formatter that attaches tenant_id and run_id
-    when set in the process environment. Lets stdout consumers (Datadog,
-    CloudWatch, Stackdriver) parse logs without ad-hoc regexes.
-    """
-
-    def format(self, record: logging.LogRecord) -> str:
-        payload = {
-            "ts": datetime.now(timezone.utc).isoformat(),
-            "level": record.levelname,
-            "logger": record.name,
-            "msg": record.getMessage(),
-        }
-        # Per-request tenant context, set by /predict handler; empty in startup.
-        tenant_id = os.environ.get("MANTIS_TENANT_ID")
-        if tenant_id:
-            payload["tenant_id"] = tenant_id
-        if record.exc_info:
-            payload["exc"] = self.formatException(record.exc_info)
-        return json.dumps(payload, default=str)
 
 
-def _configure_logging() -> None:
-    """One-time logging setup. JSON to stdout, level from LOG_LEVEL env."""
-    level_name = os.environ.get("LOG_LEVEL", "INFO").upper()
-    level = getattr(logging, level_name, logging.INFO)
-    if os.environ.get("MANTIS_LOG_FORMAT", "json").lower() == "json":
-        handler = logging.StreamHandler()
-        handler.setFormatter(_JsonLogFormatter())
-        root = logging.getLogger()
-        root.handlers[:] = [handler]
-        root.setLevel(level)
-    else:
-        logging.basicConfig(level=level)
+logger = logging.getLogger("mantis_agent.baseten_server.runtime")
 
 
-_configure_logging()
-logger = logging.getLogger("mantis_agent.baseten_server")
-
-app = FastAPI(title="Mantis CUA Baseten Workload", docs_url=None, redoc_url=None)
-
-
-SECRET_ENV_MAP = {
-    "anthropic_api_key": "ANTHROPIC_API_KEY",
-    "proxy_url": "PROXY_URL",
-    "proxy_user": "PROXY_USER",
-    "proxy_pass": "PROXY_PASS",
-    "hf_access_token": "HF_TOKEN",
-    "mantis_api_token": "MANTIS_API_TOKEN",
-}
-
-def _require_mantis_token(
-    x_mantis_token: str | None = Header(default=None, alias="X-Mantis-Token"),
-) -> TenantConfig:
-    """Container-level auth → resolved TenantConfig.
-
-    Uses a custom header (``X-Mantis-Token``) instead of ``Authorization: Bearer``
-    so it does not collide with Baseten's gateway auth, which sends
-    ``Authorization: Api-Key <baseten_key>`` to the container.
-
-    Backwards-compat: if MANTIS_TENANT_KEYS_PATH is unset and MANTIS_API_TOKEN
-    matches, returns DEFAULT_TENANT (single-tenant mode). Multi-tenant mode is
-    enabled by mounting a JSON keys file and setting MANTIS_TENANT_KEYS_PATH.
-    """
-    store = get_key_store()
-    if not store.is_multi_tenant and not os.environ.get("MANTIS_API_TOKEN", "").strip():
-        raise HTTPException(status_code=503, detail="server auth not configured")
-    if not x_mantis_token:
-        raise HTTPException(status_code=401, detail="missing X-Mantis-Token header")
-    tenant = store.resolve(x_mantis_token)
-    if tenant is None:
-        raise HTTPException(status_code=401, detail="invalid X-Mantis-Token")
-    return tenant
-
-
-def _require_run_scope(tenant: TenantConfig = Depends(_require_mantis_token)) -> TenantConfig:
-    if not tenant.has_scope("run"):
-        raise HTTPException(status_code=403, detail="tenant lacks 'run' scope")
-    return tenant
-
-
-def _read_secret(name: str) -> str:
-    path = Path("/secrets") / name
-    try:
-        return path.read_text().strip()
-    except OSError:
-        return ""
-
-
-def _resolve_anthropic_key(tenant: TenantConfig) -> str:
-    """Return the Anthropic key this tenant should use.
-
-    Reads from the secret named by the tenant's `anthropic_secret_name`
-    (each tenant can have its own Anthropic billing). Falls back to the
-    legacy ANTHROPIC_API_KEY env var if the per-tenant secret isn't present
-    on disk.
-    """
-    name = tenant.anthropic_secret_name or DEFAULT_TENANT.anthropic_secret_name
-    value = _read_secret(name)
-    if value:
-        return value
-    return os.environ.get("ANTHROPIC_API_KEY", "").strip()
-
-
-def _load_secret_environment() -> None:
-    for secret_name, env_name in SECRET_ENV_MAP.items():
-        if os.environ.get(env_name):
-            continue
-        value = _read_secret(secret_name)
-        if value:
-            os.environ[env_name] = value
-
-
-def _data_root() -> Path:
-    """Top-level data dir. Per-tenant subdirs live under this."""
-    root = Path(os.environ.get("MANTIS_DATA_DIR", "/workspace/mantis-data"))
-    root.mkdir(parents=True, exist_ok=True)
-    for child in ("results", "runs", "screenshots", "checkpoints", "chrome-profile", "tenants"):
-        (root / child).mkdir(parents=True, exist_ok=True)
-    os.environ.setdefault("MANTIS_DEBUG_DIR", str(root / "screenshots" / "claude_debug"))
-    return root
-
-
-def _tenant_root(tenant: TenantConfig) -> Path:
-    """Per-tenant subtree of the data volume. Caller cannot escape this prefix.
-
-    Layout:
-      /workspace/mantis-data/tenants/<tenant_id>/
-        ├── runs/<run_id>/{status,result,leads,events}
-        ├── checkpoints/<state_key>.json
-        ├── chrome-profile/<state_key>/
-        └── screenshots/<run_id>/
-    """
-    root = _data_root() / "tenants" / safe_state_key(tenant.tenant_id)
-    for child in ("runs", "checkpoints", "chrome-profile", "screenshots"):
-        (root / child).mkdir(parents=True, exist_ok=True)
-    return root
-
-
-def _tenant_state_key(tenant: TenantConfig, caller_state_key: str | None) -> str:
-    """Server-namespaced state key. Caller's value is sanitized + prefixed."""
-    base = safe_state_key(caller_state_key or "default")
-    return f"{safe_state_key(tenant.tenant_id)}__{base}"
-
-
-def _tenant_chrome_profile(tenant: TenantConfig, state_key: str) -> Path:
-    """Per-tenant, per-state-key Chrome profile dir."""
-    profile = _tenant_root(tenant) / "chrome-profile" / safe_state_key(state_key)
-    profile.mkdir(parents=True, exist_ok=True)
-    return profile
-
-
-def _repo_root() -> Path:
-    return Path(os.environ.get("MANTIS_REPO_ROOT", "/workspace/cua-agent"))
-
-
-_safe_state_key = safe_state_key  # backward compat alias
-
-
-_utc_now = utc_now  # backward compat alias
-
-
-def _new_run_id() -> str:
-    stamp = datetime.now(timezone.utc).strftime("%Y%m%d_%H%M%S")
-    return f"{stamp}_{uuid.uuid4().hex[:8]}"
-
-
-class _DetachedRunLogHandler(logging.Handler):
-    def __init__(self, runtime: "BasetenCUARuntime", run_id: str, thread_id: int) -> None:
-        super().__init__(level=logging.INFO)
-        self.runtime = runtime
-        self.run_id = run_id
-        self.thread_id = thread_id
-
-    def emit(self, record: logging.LogRecord) -> None:
-        if record.thread != self.thread_id:
-            return
-        try:
-            self.runtime._append_detached_event(self.run_id, self.format(record))
-        except Exception:
-            self.handleError(record)
-
-
-_parse_lead_row = parse_lead_row  # backward compat alias
-
-
-_write_leads_csv = write_leads_csv  # backward compat alias
-
-
-_plan_signature_from_steps = plan_signature_from_steps  # backward compat alias
-
-
-def _first_existing(paths: list[Path]) -> Path:
-    for path in paths:
-        if path.exists():
-            return path
-    raise FileNotFoundError("none of these paths exist: " + ", ".join(str(p) for p in paths))
-
-
-def _find_gguf(model_dir: Path, preferred: str = "") -> Path:
-    if preferred:
-        return _first_existing([Path(preferred), model_dir / preferred])
-
-    candidates = [
-        path
-        for path in model_dir.glob("*.gguf")
-        if "mmproj" not in path.name.lower()
-    ]
-    if not candidates:
-        raise FileNotFoundError(f"no model GGUF found in {model_dir}")
-
-    def rank(path: Path) -> tuple[int, str]:
-        name = path.name.lower()
-        if "q8_0" in name:
-            return (0, name)
-        if "q4_k_m" in name:
-            return (1, name)
-        return (2, name)
-
-    return sorted(candidates, key=rank)[0]
-
-
-def _find_mmproj(model_dir: Path, preferred: str = "") -> Path | None:
-    if preferred:
-        path = Path(preferred)
-        return path if path.exists() else model_dir / preferred
-    candidates = sorted(model_dir.glob("*mmproj*.gguf"))
-    return candidates[0] if candidates else None
-
-
-_wait_for_openai_server = wait_for_openai_server  # backward compat alias
-
-
-_start_local_proxy = start_local_proxy  # backward compat alias
-
-
-_build_proxy_config = build_proxy_config  # backward compat alias
+# Backwards-compat single-underscore aliases used inside the class body.
+# The class was carved out of the old single-file module where these names
+# lived alongside the runtime; rather than rewrite ~50 call sites, we keep
+# the aliases pointing at the canonical names from sibling modules.
+_data_root = data_root
+_tenant_root = tenant_root
+_tenant_state_key = tenant_state_key
+_tenant_chrome_profile = tenant_chrome_profile
+_repo_root = repo_root
+_new_run_id = new_run_id
+_find_gguf = find_gguf
+_find_mmproj = find_mmproj
+_resolve_anthropic_key = resolve_anthropic_key
+_load_secret_environment = load_secret_environment
+_read_secret = read_secret
+_DetachedRunLogHandler = DetachedRunLogHandler
+# Aliases for server_utils helpers used by name throughout the carved
+# class body. The originals live in ``server_utils``; these names are kept
+# only so the moved code does not need a global rename.
+_safe_state_key = safe_state_key
+_utc_now = utc_now
+_parse_lead_row = parse_lead_row
+_write_leads_csv = write_leads_csv
+_plan_signature_from_steps = plan_signature_from_steps
+_wait_for_openai_server = wait_for_openai_server
+_start_local_proxy = start_local_proxy
+_build_proxy_config = build_proxy_config
 
 
 class BasetenCUARuntime:
@@ -1165,388 +949,3 @@ class BasetenCUARuntime:
 runtime = BasetenCUARuntime()
 
 
-@app.on_event("startup")
-def startup() -> None:
-    runtime.load()
-
-
-@app.get("/health")
-def health() -> dict[str, Any]:
-    return {"ok": runtime.loaded, "model": runtime.model_kind}
-
-
-@app.get("/v1/health")
-def health_v1() -> dict[str, Any]:
-    """Versioned alias for /health.
-
-    The unversioned /health endpoint is what platform liveness probes target;
-    /v1/health is the same payload available under the public API path.
-    """
-    return {"ok": runtime.loaded, "model": runtime.model_kind}
-
-
-@app.get("/v1/runs/{run_id}/video")
-def get_run_video(
-    run_id: str,
-    request: Request,
-    tenant: TenantConfig = Depends(_require_mantis_token),
-) -> Any:
-    """Download the screencast for a run.
-
-    Default: serves the **polished** version (title card + captioned run
-    footage + outro card). Pass ``?raw=1`` to fetch the raw screencast
-    without overlays.
-
-    Resolves to the per-tenant run dir; returns 404 if no recording exists
-    (recording wasn't requested or ffmpeg failed). Auth requires a valid
-    token but not specifically the ``run`` scope.
-    """
-    from fastapi.responses import FileResponse
-    from mantis_agent.recorder import content_type_for
-
-    raw_only = request.query_params.get("raw", "").lower() in {"1", "true", "yes"}
-
-    safe_run_id = safe_state_key(run_id)
-    tenant_dir = _data_root() / "tenants" / safe_state_key(tenant.tenant_id)
-    runs_dir = tenant_dir / "runs" / safe_run_id
-
-    # Prefer polished by default; fall back to raw if polished is missing
-    # (e.g., ffmpeg compose failed). ?raw=1 skips polished entirely.
-    prefixes = ("recording",) if raw_only else ("recording_polished", "recording")
-    for prefix in prefixes:
-        for fmt in ("mp4", "webm", "gif"):
-            candidate = runs_dir / f"{prefix}.{fmt}"
-            if candidate.exists() and candidate.stat().st_size > 0:
-                return FileResponse(
-                    candidate,
-                    media_type=content_type_for(fmt),  # type: ignore[arg-type]
-                    filename=f"{safe_run_id}.{fmt}",
-                )
-    raise HTTPException(
-        status_code=404,
-        detail="no recording for this run "
-        "(record_video=true on /v1/predict required)",
-    )
-
-
-@app.get("/metrics")
-def metrics_endpoint() -> Any:
-    """Prometheus scrape endpoint.
-
-    Returns 503 if prometheus_client isn't installed in the container.
-    """
-    if not mantis_metrics.is_available():
-        raise HTTPException(status_code=503, detail="prometheus_client not installed")
-    return Response(
-        content=mantis_metrics.render_text(),
-        media_type=mantis_metrics.CONTENT_TYPE_LATEST,
-    )
-
-
-@app.get("/v1/models")
-def models() -> dict[str, Any]:
-    """OpenAI-compatible model listing.
-
-    Public so clients can discover the model id before sending requests.
-    Auth is enforced on the inference path itself (/v1/chat/completions).
-    """
-    return {
-        "object": "list",
-        "data": [
-            {
-                "id": runtime.model_kind,
-                "object": "model",
-                "owned_by": "mantis",
-            }
-        ],
-    }
-
-
-# Sentinel headers the upstream llama.cpp shouldn't see. We strip them so the
-# Mantis-side auth credential never reaches the inference layer.
-_PROXY_DROP_HEADERS = {
-    "host",
-    "content-length",
-    "transfer-encoding",
-    "connection",
-    "x-mantis-token",
-    "authorization",
-    "cookie",
-}
-
-
-@app.post("/v1/chat/completions")
-async def chat_completions_proxy(
-    request: Request,
-    tenant: TenantConfig = Depends(_require_run_scope),
-) -> Any:
-    """Auth-gated reverse proxy to the in-pod llama.cpp Holo3 server.
-
-    Designed for OpenAI-compat clients (the host integration's BrainHolo3 client,
-    direct `openai.OpenAI(...)` callers) that want to use Holo3 inference
-    without the full /predict orchestrator.
-
-    What this endpoint does:
-      • Validates ``X-Mantis-Token`` and resolves the tenant (must have
-        ``run`` scope).
-      • Strips Mantis-side auth headers before forwarding so the upstream
-        llama.cpp never sees them.
-      • Forwards the JSON body verbatim to the in-pod llama.cpp server at
-        ``MANTIS_LLAMA_PORT``.
-      • Passes upstream status codes and JSON bodies through.
-    """
-    body = await request.body()
-    upstream_port = os.environ.get("MANTIS_LLAMA_PORT", "18080")
-    upstream = f"http://127.0.0.1:{upstream_port}/v1/chat/completions"
-
-    headers = {
-        k: v for k, v in request.headers.items()
-        if k.lower() not in _PROXY_DROP_HEADERS
-    }
-    headers["Content-Type"] = "application/json"
-
-    logger.info(
-        "v1/chat/completions tenant=%s upstream=%s bytes=%d",
-        tenant.tenant_id,
-        upstream,
-        len(body),
-    )
-
-    try:
-        r = await run_in_threadpool(
-            requests.post,
-            upstream,
-            data=body,
-            headers=headers,
-            timeout=180,
-        )
-    except requests.RequestException as exc:
-        mantis_metrics.CHAT_COMPLETIONS.labels(
-            tenant_id=tenant.tenant_id, outcome="upstream_error"
-        ).inc()
-        logger.exception("v1/chat/completions upstream error tenant=%s", tenant.tenant_id)
-        raise HTTPException(status_code=502, detail=f"upstream error: {exc}") from exc
-
-    mantis_metrics.CHAT_COMPLETIONS.labels(
-        tenant_id=tenant.tenant_id,
-        outcome="ok" if 200 <= r.status_code < 300 else f"status_{r.status_code // 100}xx",
-    ).inc()
-
-    try:
-        payload = r.json()
-    except ValueError:
-        # Upstream returned non-JSON (rare; usually means it crashed). Surface
-        # the raw text so callers can debug.
-        return JSONResponse(
-            content={"error": {"message": r.text[:1000], "type": "upstream_error"}},
-            status_code=r.status_code if r.status_code >= 400 else 502,
-        )
-    return JSONResponse(content=payload, status_code=r.status_code)
-
-
-async def _handle_predict(
-    request: Request, tenant: TenantConfig
-) -> dict[str, Any]:
-    """Shared handler for /predict and /v1/predict.
-
-    Tier-1 + Tier-2 pipeline:
-
-    1. Pydantic validation, global cap clamp.
-    2. Per-tenant cap clamp.
-    3. State-key + Chrome-profile namespacing per tenant.
-    4. Per-tenant Anthropic key resolution.
-    5. (Tier-2) URL allowlist enforcement on the plan.
-    6. (Tier-2) Idempotency-key cache lookup.
-    7. (Tier-2) Rate-limit token consumption.
-    8. (Tier-2) Concurrency-slot acquisition (released in finally).
-    9. (Tier-2) Webhook callback registered with the runtime if requested.
-    10. Forward to the runtime.
-    """
-    try:
-        raw = await request.json()
-    except Exception as exc:
-        raise HTTPException(status_code=400, detail="request body must be JSON") from exc
-    if not isinstance(raw, dict):
-        raise HTTPException(status_code=400, detail="request body must be a JSON object")
-
-    try:
-        req = PredictRequest.model_validate(raw)
-    except Exception as exc:
-        raise HTTPException(status_code=400, detail=f"invalid request: {exc}") from exc
-
-    payload = req.model_dump(exclude_none=True)
-    payload["max_cost"] = min(
-        float(payload.get("max_cost", MAX_COST_USD)),
-        tenant.max_cost_per_run,
-    )
-    payload["max_time_minutes"] = min(
-        int(payload.get("max_time_minutes", MAX_RUNTIME_MINUTES)),
-        tenant.max_time_minutes_per_run,
-    )
-    payload["state_key"] = _tenant_state_key(tenant, payload.get("state_key"))
-
-    os.environ["ANTHROPIC_API_KEY"] = _resolve_anthropic_key(tenant)
-    os.environ["MANTIS_TENANT_ID"] = tenant.tenant_id
-    profile_dir = _tenant_chrome_profile(tenant, payload["state_key"])
-    os.environ["MANTIS_CHROME_PROFILE_DIR"] = str(profile_dir)
-
-    is_run_mode = req.action is None
-
-    # ── Tier-2: URL allowlist ────────────────────────────────────────────
-    if is_run_mode and tenant.allowed_domains:
-        plan_obj: Any = None
-        if req.task_suite is not None:
-            plan_obj = req.task_suite
-        elif req.task_file_contents:
-            try:
-                plan_obj = json.loads(req.task_file_contents)
-            except json.JSONDecodeError:
-                plan_obj = None
-        if plan_obj is not None:
-            try:
-                hosts = extract_navigate_hosts(plan_obj)
-                assert_hosts_allowed(hosts, tenant.is_domain_allowed)
-            except PermissionError as exc:
-                mantis_metrics.PREDICT_REQUESTS.labels(
-                    tenant_id=tenant.tenant_id, mode="run", outcome="denied_allowlist"
-                ).inc()
-                raise HTTPException(status_code=403, detail=str(exc)) from exc
-
-    # ── Tier-2: Idempotency-key cache ────────────────────────────────────
-    idempotency_key = request.headers.get("Idempotency-Key", "").strip()
-    if is_run_mode and idempotency_key:
-        cached = get_idempotency_cache().get(tenant.tenant_id, idempotency_key)
-        if cached is not None:
-            logger.info(
-                "predict idempotency-hit tenant=%s key=%s run_id=%s",
-                tenant.tenant_id, idempotency_key, cached.run_id,
-            )
-            mantis_metrics.PREDICT_REQUESTS.labels(
-                tenant_id=tenant.tenant_id, mode="run", outcome="idempotent_hit"
-            ).inc()
-            return cached.response
-
-    # ── Tier-2: Rate limit (token bucket) — applied to run-mode only ─────
-    limiter = get_rate_limiter()
-    if is_run_mode:
-        rate_decision = limiter.try_consume_rate_token(
-            tenant.tenant_id, tenant.rate_limit_per_minute
-        )
-        if not rate_decision.allowed:
-            mantis_metrics.RATE_LIMIT_REJECTIONS.labels(
-                tenant_id=tenant.tenant_id, kind="rate"
-            ).inc()
-            mantis_metrics.PREDICT_REQUESTS.labels(
-                tenant_id=tenant.tenant_id, mode="run", outcome="rate_limited"
-            ).inc()
-            raise HTTPException(
-                status_code=429,
-                detail=rate_decision.reason,
-                headers={"Retry-After": str(int(rate_decision.retry_after_seconds) + 1)},
-            )
-
-    # ── Tier-2: Concurrency slot ─────────────────────────────────────────
-    concurrency_acquired = False
-    if is_run_mode:
-        decision = limiter.try_acquire_concurrency_slot(
-            tenant.tenant_id, tenant.max_concurrent_runs
-        )
-        if not decision.allowed:
-            mantis_metrics.RATE_LIMIT_REJECTIONS.labels(
-                tenant_id=tenant.tenant_id, kind="concurrent"
-            ).inc()
-            mantis_metrics.PREDICT_REQUESTS.labels(
-                tenant_id=tenant.tenant_id, mode="run", outcome="rate_limited"
-            ).inc()
-            raise HTTPException(
-                status_code=429,
-                detail=decision.reason,
-                headers={"Retry-After": str(int(decision.retry_after_seconds) + 1)},
-            )
-        concurrency_acquired = True
-        mantis_metrics.CONCURRENT_RUNS.labels(tenant_id=tenant.tenant_id).set(
-            decision.concurrent
-        )
-
-    # ── Tier-2: Webhook URL — caller may override the tenant default ─────
-    webhook_url = (
-        raw.get("callback_url") or tenant.webhook_url or ""
-    ).strip()
-    if webhook_url:
-        payload["_webhook_url"] = webhook_url
-        payload["_webhook_secret_name"] = tenant.webhook_secret_name
-        payload["_tenant_id"] = tenant.tenant_id
-
-    logger.info(
-        "predict tenant=%s scope=run state_key=%s detached=%s action=%s",
-        tenant.tenant_id,
-        payload["state_key"],
-        payload.get("detached", True),
-        req.action or "run",
-    )
-
-    try:
-        response = await run_in_threadpool(runtime.run, payload)
-        mode = req.action or "run"
-        mantis_metrics.PREDICT_REQUESTS.labels(
-            tenant_id=tenant.tenant_id, mode=mode, outcome="ok"
-        ).inc()
-        if is_run_mode and idempotency_key and isinstance(response, dict):
-            get_idempotency_cache().store(
-                tenant.tenant_id, idempotency_key, response.get("run_id", ""), response
-            )
-        if is_run_mode and webhook_url and isinstance(response, dict):
-            run_id = response.get("run_id", "")
-            if response.get("status") in {"succeeded", "failed", "cancelled"}:
-                deliver_webhook_async(
-                    webhook_url,
-                    WebhookPayload(
-                        run_id=run_id,
-                        tenant_id=tenant.tenant_id,
-                        status=str(response.get("status", "")),
-                        summary=response.get("summary") or {},
-                    ),
-                    secret_name=tenant.webhook_secret_name,
-                )
-        return response
-    except ValueError as exc:
-        mantis_metrics.PREDICT_REQUESTS.labels(
-            tenant_id=tenant.tenant_id, mode="run", outcome="bad_request"
-        ).inc()
-        raise HTTPException(status_code=400, detail=str(exc)) from exc
-    except HTTPException:
-        raise
-    except Exception as exc:
-        mantis_metrics.PREDICT_REQUESTS.labels(
-            tenant_id=tenant.tenant_id, mode="run", outcome="error"
-        ).inc()
-        logger.exception("predict failed tenant=%s", tenant.tenant_id)
-        raise HTTPException(status_code=500, detail=str(exc)) from exc
-    finally:
-        if concurrency_acquired:
-            limiter.release_concurrency_slot(tenant.tenant_id)
-            mantis_metrics.CONCURRENT_RUNS.labels(tenant_id=tenant.tenant_id).set(
-                limiter.get_concurrent(tenant.tenant_id)
-            )
-
-
-@app.post("/v1/predict")
-async def predict_v1(
-    request: Request,
-    tenant: TenantConfig = Depends(_require_mantis_token),
-) -> dict[str, Any]:
-    """Tier-1 multi-tenant /predict. Validated, per-tenant capped and isolated."""
-    return await _handle_predict(request, tenant)
-
-
-@app.post("/predict")
-async def predict(
-    request: Request,
-    tenant: TenantConfig = Depends(_require_mantis_token),
-) -> dict[str, Any]:
-    """Backwards-compat alias for /v1/predict.
-
-    Kept indefinitely for callers built against the v1.0 deployment shape.
-    Identical behavior to /v1/predict.
-    """
-    return await _handle_predict(request, tenant)
